search pages container component 분리
