+++ src/components/search/DetailComponent.tsx
@@ -1,0 +1,5 @@
+import DetailContainer from "../../containers/search/DetailContainer";
+
+export default function DetailComponent(): JSX.Element {
+  return <DetailContainer />;
+}
