+++ src/components/search/IndexComponent.tsx
@@ -1,0 +1,5 @@
+import IndexContainer from "../../containers/search/IndexContainer";
+
+export default function IndexComponent(): JSX.Element {
+  return <IndexContainer />;
+}
